% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longArray.R
\name{longArray}
\alias{longArray}
\title{construct a longArray instance}
\usage{
longArray(row.ids, col.ids, data, ...)
}
\arguments{
\item{row.ids}{define}

\item{col.ids}{define}

\item{data}{define}

\item{\dots}{define}
}
\description{
construct a longArray instance
}
