% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viabHeatmap.R
\name{viabDots}
\alias{viabDots}
\title{depict data from a 2-drug screen}
\usage{
viabDots(dat, drugA = "5-FU", drugB = "ABT-888", incell_line = "A2058")
}
\arguments{
\item{dat}{data.frame with specific fields available}

\item{drugA}{character(1) drug name}

\item{drugB}{character(1) drug name}

\item{incell_line}{character(1) cell_line name}
}
\value{
a ggplot object
}
\description{
depict data from a 2-drug screen
}
\examples{
data(dc)
viabDots(dc)
}
