% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viabReg.R
\name{viabReg}
\alias{viabReg}
\title{model viability results from a 2-drug screen}
\usage{
viabReg(dat, drugA = "5-FU", drugB = "ABT-888", incell_line = "A2058")
}
\arguments{
\item{dat}{data.frame with specific fields available}

\item{drugA}{character(1) drug name}

\item{drugB}{character(1) drug name}

\item{incell_line}{character(1) cell_line name}
}
\value{
lm of the model for viability with concA x concB interaction and all main effects
}
\description{
model viability results from a 2-drug screen
}
\note{
Assumes independence of all replicates.
}
\examples{
data(dc)
viabReg(dc)
}
